Lutero branch

+++ includes/users-processor.inc.php
@@ -1,21 +1,30 @@
 <?php
 
-if (isset($_POST["delete-user-submit"])) {
+// if (isset($_POST["delete-user-submit"])) {
 
 require "dbh.inc.php";
-$userId = $_POST['userId'];
+$userId = $_POST['id'];
 
-$sql = "DELETE FROM users WHERE idUsers='$userId'";
-if ($conn->query($sql) === TRUE) {
-header("Location: ../users.php?user-removed=success");
+$sql = "SELECT * FROM users WHERE idUsers='$userId'";
+$result = mysqli_query($conn, $sql);
+$totalrows = mysqli_num_rows($result);
+
+if ($totalrows > 0) {
+    $sql = "DELETE FROM users WHERE idUsers='$userId'";
+    if ($conn->query($sql) === TRUE) {
+        echo "1";
+        exit();
+    }
+    echo "0";
+    exit();
 } else {
-header("Location: ../users.php?user-removed=error");
+    echo "0";
+    exit();
 }
-$userRemovedSuccessAlert;
-exit();
+
 mysqli_stmt_close($stmt);
 mysqli_close($conn);
 
-}
+// }
 
 // TODO Create a promotion button in users.php
